@@ -87,7 +119,7 @@
   "name": "react-addons-css-transition-group",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^15.2.1"
+    "react": "^15.4.2"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
