@@ -3,59 +3,82 @@
     [
       {
         "name": "react-addons-shallow-compare",
-        "raw": "react-addons-shallow-compare@15.2.1",
-        "rawSpec": "15.2.1",
+        "raw": "react-addons-shallow-compare@15.5.2",
+        "rawSpec": "15.5.2",
         "scope": null,
-        "spec": "15.2.1",
+        "spec": "15.5.2",
         "type": "version"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend"
     ]
   ],
-  "_from": "react-addons-shallow-compare@15.2.1",
-  "_id": "react-addons-shallow-compare@15.2.1",
+  "_from": "react-addons-shallow-compare@15.5.2",
+  "_id": "react-addons-shallow-compare@15.5.2",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-addons-shallow-compare",
-  "_nodeVersion": "4.4.5",
+  "_nodeVersion": "7.7.4",
   "_npmOperationalInternal": {
-    "host": "packages-16-east.internal.npmjs.com",
-    "tmp": "tmp/react-addons-shallow-compare-15.2.1.tgz_1468017056920_0.44626194052398205"
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/react-addons-shallow-compare-15.5.2.tgz_1491708170167_0.5065773921087384"
   },
   "_npmUser": {
-    "email": "paul@oshannessy.com",
-    "name": "zpao"
+    "email": "dan.abramov@gmail.com",
+    "name": "gaearon"
   },
-  "_npmVersion": "2.15.5",
-  "_phantomChildren": {},
+  "_npmVersion": "4.1.2",
+  "_phantomChildren": {
+    "isomorphic-fetch": "2.2.1",
+    "loose-envify": "1.2.0",
+    "object-assign": "4.1.0",
+    "promise": "7.1.1",
+    "setimmediate": "1.0.5",
+    "ua-parser-js": "0.7.12"
+  },
   "_requested": {
     "name": "react-addons-shallow-compare",
-    "raw": "react-addons-shallow-compare@15.2.1",
-    "rawSpec": "15.2.1",
+    "raw": "react-addons-shallow-compare@15.5.2",
+    "rawSpec": "15.5.2",
     "scope": null,
-    "spec": "15.2.1",
+    "spec": "15.5.2",
     "type": "version"
   },
   "_requiredBy": [
     "/",
     "/pui-react-mixins"
   ],
-  "_resolved": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.2.1.tgz",
-  "_shasum": "b7b6b52e94a3f7c5185c2d4a258eb7e9cd7e963a",
+  "_resolved": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.5.2.tgz",
+  "_shasum": "7cb0ee7acc8d7c93fcc202df0bf47ba916a7bdad",
   "_shrinkwrap": null,
-  "_spec": "react-addons-shallow-compare@15.2.1",
+  "_spec": "react-addons-shallow-compare@15.5.2",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend",
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
   },
-  "dependencies": {},
-  "description": "This package provides the React shallowCompare add-on.",
-  "devDependencies": {},
+  "dependencies": {
+    "fbjs": "^0.8.4",
+    "object-assign": "^4.1.0"
+  },
+  "description": ">**Note:** >This is a legacy React addon, and is no longer maintained. > >We don't encourage using it in new code, but it exists for backwards compatibility.   >The recommended migration path is to use [`React.PureComponent`](https://facebook.github.io/re",
+  "devDependencies": {
+    "jest": "^19.0.2",
+    "react": "^15.4.2",
+    "react-addons-test-utils": "^15.4.2",
+    "react-dom": "^15.4.2"
+  },
   "directories": {},
   "dist": {
-    "shasum": "b7b6b52e94a3f7c5185c2d4a258eb7e9cd7e963a",
-    "tarball": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.2.1.tgz"
+    "shasum": "7cb0ee7acc8d7c93fcc202df0bf47ba916a7bdad",
+    "tarball": "https://registry.npmjs.org/react-addons-shallow-compare/-/react-addons-shallow-compare-15.5.2.tgz"
   },
+  "files": [
+    "LICENSE",
+    "PATENTS",
+    "README.md",
+    "index.js",
+    "react-addons-shallow-compare.js",
+    "react-addons-shallow-compare.min.js"
+  ],
   "homepage": "https://github.com/facebook/react#readme",
   "keywords": [
     "react",
