@@ -87,7 +134,7 @@
   "name": "react-dom",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^15.2.1"
+    "react": "^15.5.4"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
