@@ -95,28 +97,27 @@
     "gzip-size": "^3.0.0",
     "karma": "^1.1.2",
     "karma-browserstack-launcher": "^1.0.1",
-    "karma-chrome-launcher": "^1.0.1",
+    "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.1.1",
     "karma-mocha": "^1.1.1",
     "karma-mocha-reporter": "^2.0.5",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.7.0",
-    "mocha": "^2.5.3",
-    "pretty-bytes": "^3.0.1",
+    "karma-webpack": "^2.0.3",
+    "mocha": "^3.2.0",
+    "pretty-bytes": "^4.0.2",
     "qs": "^6.2.1",
-    "react": "^15.3.0",
-    "react-addons-css-transition-group": "^15.3.0",
-    "react-addons-test-utils": "^15.3.0",
-    "react-dom": "^15.3.0",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
+    "react-transition-group": "^1.1.1",
     "rimraf": "^2.5.4",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.16.1",
     "webpack": "^1.13.1",
     "webpack-dev-middleware": "^1.6.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "5a19156678810e01d81901f9c0fef63284b8a514",
-    "tarball": "https://registry.npmjs.org/react-router/-/react-router-3.0.2.tgz"
+    "shasum": "c3b7873758045a8bbc9562aef4ff4bc8cce7c136",
+    "tarball": "https://registry.npmjs.org/react-router/-/react-router-3.0.5.tgz"
   },
   "files": [
     "*.md",
