@@ -3,17 +3,11 @@ var _extends = Object.assign || function (target) { for (var i = 1; i < argument
 function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
 
 import React from 'react';
+import createReactClass from 'create-react-class';
+import { bool, object, string, func, oneOfType } from 'prop-types';
 import invariant from 'invariant';
 import { routerShape } from './PropTypes';
 import { ContextSubscriber } from './ContextUtils';
-
-var _React$PropTypes = React.PropTypes,
-    bool = _React$PropTypes.bool,
-    object = _React$PropTypes.object,
-    string = _React$PropTypes.string,
-    func = _React$PropTypes.func,
-    oneOfType = _React$PropTypes.oneOfType;
-
 
 function isLeftClickEvent(event) {
   return event.button === 0;
@@ -46,15 +40,9 @@ function resolveToLocation(to, router) {
  * You could use the following component to link to that route:
  *
  *   <Link to={`/posts/${post.id}`} />
- *
- * Links may pass along location state and/or query string parameters
- * in the state/query props, respectively.
- *
- *   <Link ... query={{ show: true }} state={{ the: 'state' }} />
  */
-var Link = React.createClass({
+var Link = createReactClass({
   displayName: 'Link',
-
 
   mixins: [ContextSubscriber('router')],
 
@@ -64,9 +52,6 @@ var Link = React.createClass({
 
   propTypes: {
     to: oneOfType([string, object, func]),
-    query: object,
-    hash: string,
-    state: object,
     activeStyle: object,
     activeClassName: string,
     onlyActiveOnIndex: bool.isRequired,
