@@ -1,12 +1,9 @@
-import React from 'react';
+import createReactClass from 'create-react-class';
+import { string, object } from 'prop-types';
 import invariant from 'invariant';
 import { createRouteFromReactElement as _createRouteFromReactElement } from './RouteUtils';
 import { formatPattern } from './PatternUtils';
 import { falsy } from './InternalPropTypes';
-
-var _React$PropTypes = React.PropTypes,
-    string = _React$PropTypes.string,
-    object = _React$PropTypes.object;
 
 /**
  * A <Redirect> is used to declare another URL path a client should
@@ -16,10 +13,8 @@ var _React$PropTypes = React.PropTypes,
  * and are traversed in the same manner.
  */
 /* eslint-disable react/require-render-return */
-
-var Redirect = React.createClass({
+var Redirect = createReactClass({
   displayName: 'Redirect',
-
 
   statics: {
     createRouteFromReactElement: function createRouteFromReactElement(element) {
